ZZALIB: 1.0.27 Enhancement: Change POST FORM to accept null pointer for request stream. Code tidy: xml_reader - not using result of ios::good(), double to long conversion ZZALOG: 3.1.108 Enhancement: make adi_writer::item_to_adif a statis to allow other classes to use it. Enhancement: Add upload_single_qso() to eQSL, LotW and Clublog handlers. Call this when importing from auto-import or datagram Code tidy: wsjtx_handler - double to int conversion Code tid: dxa_if - double to float conversion.

+++ zzalog/dxa_if.cpp
@@ -1794,7 +1794,7 @@ void dxa_if::zoom_azimuthal() {
 	switch (centre_mode_) {
 	case HOME:
 		// Zoom in to just the displayed QSOs
-		zoom = (float)((EARTH_RADIUS * PI) / (float)furthest_) * 0.95;
+		zoom = (float)((EARTH_RADIUS * PI) / furthest_ * 0.95);
 		map->PutZoom(zoom);
 		break;
 	case ZERO:
